% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addrPhoneExtractor.R
\name{addrPhoneExtractor}
\alias{addrPhoneExtractor}
\title{提供文本中的电话及地址信息}
\usage{

  addrPhoneExtractor(file = "~/pkgs/tsda/data-raw/src_data/addrPhone_data.xlsx")
}
\arguments{
\item{file}{录入Excel文件名}
}
\value{
返回值
}
\description{
提供文本中的电话及地址信息
}
\examples{
addrPhoneExtrator();
}
