% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/21-allocation.R
\name{allocate}
\alias{allocate}
\title{在2个数据框中进行自上而下分配运算}
\usage{
allocate(df1, df2)
}
\arguments{
\item{df1}{数据1}

\item{df2}{数据2}
}
\value{
取最小行返回数据框
}
\description{
在2个数据框中进行自上而下分配运算
}
\examples{
res <- allocate(data.frame(a=letters),data.frame(b=10:1));
}
