% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_textmining.R
\name{book_tidy}
\alias{book_tidy}
\title{tidy books}
\usage{
book_tidy(books)
}
\arguments{
\item{books}{get from gutenberg_download}
}
\value{
tibble result
}
\description{
tidy books
}
\examples{
book_tidy()
}
