% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMYK.R
\name{CMYK}
\alias{CMYK}
\title{用于输入一个完成的CYMK资料}
\usage{
CMYK(data)
}
\arguments{
\item{data}{输入数据，列表或data.frame,提供C,Y,M,K四个参数}
}
\value{
返回值
}
\description{
用于输入一个完成的CYMK资料
}
\examples{
CYMK(list(C=1:100,Y=1:100,M=1:100,K=1:100));
}
