% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-color.R
\name{CMYK2RGB}
\alias{CMYK2RGB}
\title{用于将CMYK模式的图片设置成RGB格式}
\usage{
CMYK2RGB(C = 0, M = 0, Y = 0, K = 0)
}
\arguments{
\item{C}{青色CyanW值，0-100之间;}

\item{M}{品红色Magenta值，0-100之间；}

\item{Y}{黄色Yelloww值，0-100之间；}

\item{K}{Black or Key Board 值，0-100之间}
}
\value{
返回列表
}
\description{
用于将CMYK模式的图片设置成RGB格式
}
\examples{
CMYK2RGB(1,0,0,0);
}
