% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-vec_character.R
\name{combine_rows_into_longColumn}
\alias{combine_rows_into_longColumn}
\title{合并多行数据}
\usage{
combine_rows_into_longColumn(data, id_var, value_var)
}
\arguments{
\item{data}{数据框}

\item{id_var}{int类型的内码}

\item{value_var}{变量}
}
\value{
返回值
}
\description{
合并多行数据
}
\examples{
combine_rows_into_longColumn();
}
