% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-df_rowOpers.R
\name{comboLogi_to_r}
\alias{comboLogi_to_r}
\title{针对逻辑值转化为符合R语言规范的规格}
\usage{
comboLogi_to_r(x)
}
\arguments{
\item{x}{数据}
}
\value{
返回值
}
\description{
针对逻辑值转化为符合R语言规范的规格
}
\examples{
comboLogi_to_r();
}
