% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-df_rowOpers.R
\name{comboLogi_to_rs}
\alias{comboLogi_to_rs}
\title{将and or转化为R语言中的转换符}
\usage{
comboLogi_to_rs(x)
}
\arguments{
\item{x}{充满and or的字符串}
}
\value{
返回R中的比较符
}
\description{
将and or转化为R语言中的转换符
}
\examples{
comboLogi_to_rs();
}
