% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05a_date.R
\name{date_add}
\alias{date_add}
\title{处理日期数据}
\usage{
date_add(startDate = "1984-01-01", length = 5)
}
\arguments{
\item{startDate}{开始日期}

\item{length}{个数}
}
\value{
返回值
}
\description{
处理日期数据
}
\examples{
date_add();
}
