% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-datetime.R
\name{date_diff}
\alias{date_diff}
\title{计算2个日期之间的差异}
\usage{
date_diff(startDate, endDate, unit = "d", hourPerDay = 8,
  is.hr = TRUE)
}
\arguments{
\item{startDate}{开始日期}

\item{endDate}{结束日期}

\item{unit}{日期差异的单位,默认为d}

\item{hourPerDay}{每天工作小时数}

\item{is.hr}{是否HR计时}
}
\value{
返回值
}
\description{
计算2个日期之间的差异
}
\examples{
date_diff('2019-6-11','2019-07-22')
}
