% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05a_date.R
\name{date_minus}
\alias{date_minus}
\title{生成日期序列}
\usage{
date_minus(startDate = "1984-01-01", endDate = "1984-01-05")
}
\arguments{
\item{startDate}{开始日志}

\item{endDate}{结束日期}
}
\value{
返回值
}
\description{
生成日期序列
}
\examples{
date_minus();
}
