% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-datetime.R
\name{datetime_diff}
\alias{datetime_diff}
\title{计算出两个datetime类型数据的时间差}
\usage{
datetime_diff(x, y, unit = "h", digit = 1)
}
\arguments{
\item{x}{第一个数据}

\item{y}{第一个数据}

\item{unit}{时间差单位，默认为h表示小时}

\item{digit}{保留小数位数}
}
\value{
返回值
}
\description{
计算出两个datetime类型数据的时间差
}
\examples{
datetime_diff(as.datetime('2019-08-14 08:14:00'),as.datetime('2019-08-15 08:14:00'))
}
