% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-data_frame.R
\name{df_addCol}
\alias{df_addCol}
\title{数据框添加列}
\usage{
df_addCol(data = iris, col_name = "bbc", col_value = "123")
}
\arguments{
\item{data}{数据框}

\item{col_name}{列名}

\item{col_value}{列值}
}
\value{
返回值
}
\description{
数据框添加列
}
\examples{
df_addCol()
}
