% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-data_frame.R
\name{df_as_character}
\alias{df_as_character}
\title{将数据框中的所有列作为文本处理}
\usage{
df_as_character(df)
}
\arguments{
\item{df}{处理框}
}
\value{
返回值
}
\description{
将数据框中的所有列作为文本处理
}
\examples{
df_as_character();
}
