% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-data_frame.R
\name{df_as_groupList}
\alias{df_as_groupList}
\title{将数据框转换成分组列表}
\usage{
df_as_groupList(df, group_col, value_col)
}
\arguments{
\item{df}{数据框}

\item{group_col}{分组列名}

\item{value_col}{值名}
}
\value{
返回二维list
}
\description{
将数据框转换成分组列表
}
\examples{
df_as_groupList()
}
