% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-data_frame.R
\name{df_blackList_tagging}
\alias{df_blackList_tagging}
\title{针对数据框进行打款}
\usage{
df_blackList_tagging(
  data,
  var_text = "Ftxt",
  dict,
  var_flag = "Fflag",
  var_tag_by = "Ftag_by"
)
}
\arguments{
\item{data}{数据框}

\item{var_text}{字符}

\item{dict}{规则表}

\item{var_flag}{是否打中}

\item{var_tag_by}{判断标记}
}
\value{
返回数据框
}
\description{
针对数据框进行打款
}
\examples{
df_blackList_tagging()
}
