% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-data_frame.R
\name{df_combineRows}
\alias{df_combineRows}
\title{将数据按标记进行处理}
\usage{
df_combineRows(data, var_txt, var_flag)
}
\arguments{
\item{data}{数据}

\item{var_txt}{内容}

\item{var_flag}{标记}
}
\value{
返回值
}
\description{
将数据按标记进行处理
}
\examples{
df_combineRows()
}
