% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-df_summaise_tableLevel.R
\name{df_count}
\alias{df_count}
\title{进行表级的计数，可以多列一起进行计数}
\usage{
df_count(data, fieldName = c("mpg", "cyl", "disp"))
}
\arguments{
\item{data}{数据框}

\item{fieldName}{字名名}
}
\value{
返回值
}
\description{
进行表级的计数，可以多列一起进行计数
}
\examples{
df_count();
df_count(data=mtcars,
fieldName=c('mpg','cyl','disp')
);
}
