% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-df_rowOpers.R
\name{df_filter}
\alias{df_filter}
\title{将数据集进行过滤}
\usage{
df_filter(
  data = iris,
  fieldName = c("Species", "Sepal.Width"),
  comparerSign = c("=", ">="),
  value_vec = c("'setosa'", "3.0"),
  comboLogi_vec = c("and", "and")
)
}
\arguments{
\item{data}{数据框}

\item{fieldName}{过滤字段}

\item{comparerSign}{比例符号}

\item{value_vec}{值设置}

\item{comboLogi_vec}{逻辑连接条件}
}
\value{
返回值
}
\description{
将数据集进行过滤
}
\examples{
df_filter();
df_filter(iris,fieldName = 'Sepal.Width',comparerSign = '=',value_vec = '3.5',comboLogi_vec = 'and');
}
