% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-df_detail_groupLevel.R
\name{df_groupBy_topN}
\alias{df_groupBy_topN}
\title{获取数据框每个分组前N个数据}
\usage{
df_groupBy_topN(data = iris, gp_fieldName = c("Species",
  "Petal.Width"), order_fieldName = "Sepal.Width", N = 3)
}
\arguments{
\item{data}{数据框}

\item{gp_fieldName}{分组字段，支持组合}

\item{order_fieldName}{排序字段}

\item{N}{前N个数据}
}
\value{
返回数据框
}
\description{
获取数据框每个分组前N个数据
}
\examples{
 df_groupBy_topN();
df_groupBy_topN(data = iris,gp_fieldName = 'Species',order_fieldName = 'Sepal.Width',N = 2

);
}
