% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-df_joinOpers.R
\name{df_innerJoin_bySameColNames}
\alias{df_innerJoin_bySameColNames}
\title{将两个数据框进行内连接，连接字段名相同}
\usage{
df_innerJoin_bySameColNames(data_frame1, data_fram2, byFieldNames)
}
\arguments{
\item{data_frame1}{第1数据框}

\item{data_fram2}{第2数据框}

\item{byFieldNames}{连接字段名}
}
\value{
返回值
}
\description{
将两个数据框进行内连接，连接字段名相同
}
\examples{
library(tsdo);
data("joinData1");
data("joinData2");
filed_by <-c('fname','fclass');
df_innerJoin_bySameColNames(joinData1,joinData2,filed_by) 
}
