% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-data_frame.R
\name{df_left}
\alias{df_left}
\title{get the datarightpart}
\usage{
df_left(data, var_txt, var_right, var_left)
}
\arguments{
\item{data}{data}

\item{var_txt}{txt}

\item{var_right}{right}

\item{var_left}{left}
}
\value{
return
}
\description{
get the datarightpart
}
\examples{
df_left
}
