% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-df_joinOpers.R
\name{df_leftJoin_byDiffColNames}
\alias{df_leftJoin_byDiffColNames}
\title{两个数据框左连接，列名不同，但一一对应}
\usage{
df_leftJoin_byDiffColNames(data_frame1, data_frame2, leftBy_fieldName,
  rightBy_fieldName)
}
\arguments{
\item{data_frame1}{第1数据框}

\item{data_frame2}{每2数据框}

\item{leftBy_fieldName}{左列名}

\item{rightBy_fieldName}{右列表}
}
\value{
返回值
}
\description{
两个数据框左连接，列名不同，但一一对应
}
\examples{
library(tsdo);
data("joinData1");
data("joinData3");
left_by <-c('fname','fclass');
right_by <-c('titileName','titleClass');
df_leftJoin_byDiffColNames(joinData1,joinData3,left_by,right_by);
}
