% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-data_frame.R
\name{df_mergeColumn}
\alias{df_mergeColumn}
\title{合并data_frame的列信息，用于文本合并}
\usage{
df_mergeColumn(data_frame)
}
\arguments{
\item{data_frame}{数据框}
}
\value{
返回值
}
\description{
合并data_frame的列信息，用于文本合并
}
\examples{
df_mergeColumn();
}
