% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-data_frame.R
\name{df_orderData}
\alias{df_orderData}
\title{增加对任意数据框进行排序功能}
\usage{
df_orderData(srcData = "data", order_by, is_asc = order_d)
}
\arguments{
\item{srcData}{源数据框}

\item{order_by}{排序字段，向量}

\item{is_asc}{逻辑变更，默认T为升序}
}
\value{
返回值
}
\description{
增加对任意数据框进行排序功能
}
\examples{
df_orderData()
}
