% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-df_rowOpers.R
\name{df_paging}
\alias{df_paging}
\title{针对任意的数据框进行分页处理}
\usage{
df_paging(data = iris, each_page = 7000L, skip_row = 0L)
}
\arguments{
\item{data}{数据}

\item{each_page}{每页数量}

\item{skip_row}{跳过行数}
}
\value{
返回值
}
\description{
针对任意的数据框进行分页处理
}
\examples{
df_paging()
}
