% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-data_frame.R
\name{df_right}
\alias{df_right}
\title{提供数据框的右边部分}
\usage{
df_right(data, var_txt, var_left, var_right)
}
\arguments{
\item{data}{数据}

\item{var_txt}{文本内容}

\item{var_left}{左边已有内容}

\item{var_right}{待生成右边内容}
}
\value{
返回值
}
\description{
提供数据框的右边部分
}
\examples{
df_right()
}
