% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-data_frame.R
\name{df_rowColEx}
\alias{df_rowColEx}
\title{将数据框的行与列进行转化，}
\usage{
df_rowColEx(data)
}
\arguments{
\item{data}{数据框，要求第一列必须为文本，后续用于形成列标题}
}
\value{
返回转化后的数据框
}
\description{
将数据框的行与列进行转化，
}
\examples{
rowColEx(data);
}
