% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-data_frame.R
\name{df_rowRepMulti}
\alias{df_rowRepMulti}
\title{将具有一行数据的数据框复制为多行}
\usage{
df_rowRepMulti(data, times = 6)
}
\arguments{
\item{data}{原始数据框}

\item{times}{复制行数}
}
\value{
返回值
}
\description{
将具有一行数据的数据框复制为多行
}
\examples{
df_rowRepMulti();
}
