% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-data_frame.R
\name{df_rowRepMultiByDate}
\alias{df_rowRepMultiByDate}
\title{根据日期及时间数据拆分笔}
\usage{
df_rowRepMultiByDate(data, startDateName = "FStartDate",
  endDateName = "FEndDate")
}
\arguments{
\item{data}{数据}

\item{startDateName}{开始日期字段}

\item{endDateName}{结束日期字段}
}
\value{
返回值
}
\description{
根据日期及时间数据拆分笔
}
\examples{
df_rowRepMultiByDate();
}
