% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-data_frame.R
\name{df_selectCol}
\alias{df_selectCol}
\title{数据框选取列}
\usage{
df_selectCol(data, colSelected)
}
\arguments{
\item{data}{数据框}

\item{colSelected}{列名}
}
\value{
返回值
}
\description{
数据框选取列
}
\examples{
df_selectCol();
}
