% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-data_frame.R
\name{df_setColCaption}
\alias{df_setColCaption}
\title{设置列名}
\usage{
df_setColCaption(data, col_names, caption_names)
}
\arguments{
\item{data}{数据框}

\item{col_names}{列名}

\item{caption_names}{列标称名}
}
\value{
返回数据框
}
\description{
设置列名
}
\examples{
df_setColCaption()
}
