% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-df_setsOpers.R
\name{df_setdiff}
\alias{df_setdiff}
\title{将data存在于data_frame1而不存在于data_frame2的记录取出}
\usage{
df_setdiff(data_frame1, data_frame2)
}
\arguments{
\item{data_frame1}{第一个数据框}

\item{data_frame2}{第二个数据框}
}
\value{
返回值
}
\description{
将data存在于data_frame1而不存在于data_frame2的记录取出
}
\examples{
library(tsdo);
data('setExample1')
data('setExample2')
df_setdiff(setExample1,setExample2);
}
