% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-data_frame.R
\name{df_setLabel}
\alias{df_setLabel}
\title{针对数据框设置标准}
\usage{
df_setLabel(
  data,
  var_txt = "FLog",
  keyword = "捷豹路虎官方旗舰店",
  var_flag = "FIsCsp"
)
}
\arguments{
\item{data}{数据框}

\item{var_txt}{内容字段}

\item{keyword}{关键词}

\item{var_flag}{标记字段}
}
\value{
返回一个数据框
}
\description{
针对数据框设置标准
}
\examples{
df_setLabel
}
