% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-data_frame.R
\name{df_setList_char}
\alias{df_setList_char}
\title{将列如果是列表转移成字符串}
\usage{
df_setList_char(data)
}
\arguments{
\item{data}{数据框}
}
\value{
返回值
}
\description{
将列如果是列表转移成字符串
}
\examples{
df_setList_char()
}
