% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-data_frame.R
\name{df_splitByCol}
\alias{df_splitByCol}
\title{针对数据分列处理，按指定列进行处理}
\usage{
df_splitByCol(
  data,
  var_txt = "FLog",
  var_split = "log_datetime",
  var_left = "author",
  left_skip = 2,
  var_right = "content",
  right_skip = 4
)
}
\arguments{
\item{data}{数据框}

\item{var_txt}{原始内容列}

\item{var_split}{分列依据字段}

\item{var_left}{左边字段}

\item{left_skip}{分列左边忽略字符数}

\item{var_right}{右边字段}

\item{right_skip}{分列右边忽略字符数}
}
\value{
返回数据框
}
\description{
针对数据分列处理，按指定列进行处理
}
\examples{
df_splitByCol
}
