% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-df_summaise_tableLevel.R
\name{df_summarise_allCols_oneFun}
\alias{df_summarise_allCols_oneFun}
\title{针对数据框所有列使用同一个函数处理}
\usage{
df_summarise_allCols_oneFun(data, fun)
}
\arguments{
\item{data}{}

\item{fun}{函数名，只指定一个}
}
\value{
返回值
}
\description{
针对数据框所有列使用同一个函数处理
}
\examples{
 df_summarise_allCols_oneFun();
}
