% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-df_summaise_tableLevel.R
\name{df_summarise_cols_funs}
\alias{df_summarise_cols_funs}
\title{针对表中的字段进行统计，在不分的情况下}
\usage{
df_summarise_cols_funs(data = mtcars, fieldName = c("mpg", "cyl",
  "disp"), funName = c("mean", "sum", "length"),
  fieldCaption = letters[1:3])
}
\arguments{
\item{data}{数据集}

\item{fieldName}{字段名称}

\item{funName}{函数名称}

\item{fieldCaption}{字段标题}
}
\value{
返回值
}
\description{
针对表中的字段进行统计，在不分的情况下
}
\examples{
df_summarise_cols_funs();
df_summarise_cols_funs(data=mtcars,
fieldName=c('mpg','cyl','disp'),
funName=c('mean','sum','length'),
fieldCaption=letters[1:3]
);
}
