% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-df_summaise_tableLevel.R
\name{df_summarise_cols_oneFun}
\alias{df_summarise_cols_oneFun}
\title{表的指定列使用同一个函数处理}
\usage{
df_summarise_cols_oneFun(
  data = mtcars,
  fieldName = c("mpg", "cyl", "disp"),
  fun
)
}
\arguments{
\item{data}{数据框}

\item{fieldName}{字段名}

\item{fun}{函数}
}
\value{
返回值
}
\description{
表的指定列使用同一个函数处理
}
\examples{
df_summarise_cols_oneFun();
df_summarise_cols_oneFun(data=mtcars,
fieldName=c('mpg','cyl'),length);
}
