% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/32-env.R
\name{env_ls}
\alias{env_ls}
\title{show all the object in a env}
\usage{
env_ls(env)
}
\arguments{
\item{env}{environment}
}
\value{
return object vect
}
\description{
show all the object in a env
}
\examples{
env_ls()
}
