% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/32-env.R
\name{envs_are_identical}
\alias{envs_are_identical}
\title{assert envs are the identical}
\usage{
envs_are_identical(env1, env2)
}
\arguments{
\item{env1}{env1}

\item{env2}{env2}
}
\value{
return value
}
\description{
assert envs are the identical
}
\examples{
envs_are_identical()
}
