% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-vec_logical.R
\name{\%||\%}
\alias{\%||\%}
\title{定义逻辑判断方案，选择第一个方案或备用方案}
\usage{
a \%||\% b
}
\arguments{
\item{a}{第一个方案}

\item{b}{备用方案}
}
\value{
返回值
}
\description{
定义逻辑判断方案，选择第一个方案或备用方案
}
\examples{
'a' \%||\% 'b'
}
