% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-vec_character.R
\name{ifnull}
\alias{ifnull}
\title{设置文本为null值为相应的替代值}
\usage{
ifnull(x, value = "")
}
\arguments{
\item{x}{原始向量}

\item{value}{替代的数据}
}
\value{
返回值
}
\description{
设置文本为null值为相应的替代值
}
\examples{
ifnull();
}
