% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-datetime.R
\name{is.Am}
\alias{is.Am}
\title{判断时间字段是否为上午}
\usage{
is.Am(x, AmStartTime = "6:00:00", AmEndTime = "12:00:00")
}
\arguments{
\item{x}{时间字段}

\item{AmEndTime}{上午结束时间}
}
\value{
返回值
}
\description{
判断时间字段是否为上午
}
\examples{
is.Am(9:00:00);
}
