% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-datetime.R
\name{is.datetime}
\alias{is.datetime}
\title{判断对象是否为日期时间型}
\usage{
is.datetime(x)
}
\arguments{
\item{x}{对象}
}
\value{
返回值
}
\description{
判断对象是否为日期时间型
}
\examples{
is.datetime(as.datetime('2019-08-14 08:14:00'));
}
