% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-datetime.R
\name{is.Morning}
\alias{is.Morning}
\title{判断时间是否为凌晨}
\usage{
is.Morning(x, morningStartTime = "0:00:00", morningEndTime = "6:00:00")
}
\arguments{
\item{x}{时间}

\item{morningStartTime}{凌晨开始时间}

\item{morningEndTime}{凌晨结束时间}
}
\value{
返回值
}
\description{
判断时间是否为凌晨
}
\examples{
is.Morning();
}
