% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-datetime.R
\name{is.Night}
\alias{is.Night}
\title{判断是否为夜晚}
\usage{
is.Night(x, nightStartTime = "18:30:00")
}
\arguments{
\item{x}{数值}

\item{nightStartTime}{夜晚开始时间}
}
\value{
返回值
}
\description{
判断是否为夜晚
}
\examples{
is.Night('19:00:00');
}
