% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-datetime.R
\name{is.Pm}
\alias{is.Pm}
\title{判断时间是否为下午时间}
\usage{
is.Pm(x, startTime = "12:00:00", endTime = "18:30:00")
}
\arguments{
\item{x}{日期范围}

\item{startTime}{下午开始时间}

\item{endTime}{下午结束时间}
}
\value{
返回值
}
\description{
判断时间是否为下午时间
}
\examples{
is.Pm('14:00:00');
}
