% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-list.R
\name{list_init}
\alias{list_init}
\title{处理初始化list操作}
\usage{
list_init(len = 3, listNames = NULL)
}
\arguments{
\item{len}{元素个数}

\item{listNames}{元素名称}
}
\value{
返回值
}
\description{
处理初始化list操作
}
\examples{
library(tsdo);
aa <-list_init(3);
aa;
bb<-list_init(listNames = LETTERS);
bb;

}
