% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-vec_character.R
\name{na_value}
\alias{na_value}
\title{针对空值进行替代}
\usage{
na_value(test_value, replace_value)
}
\arguments{
\item{test_value}{测试值}

\item{replace_value}{替代值}
}
\value{
返回值
}
\description{
针对空值进行替代
}
\examples{
na_value();
}
