% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-vec_character.R
\name{na_values}
\alias{na_values}
\title{针对数据进行处理}
\usage{
na_values(test_values, replace_values)
}
\arguments{
\item{test_values}{测试值}

\item{replace_values}{替代值}
}
\value{
返回值
}
\description{
针对数据进行处理
}
\examples{
na_values();
}
