% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_textmining.R
\name{ngram_count}
\alias{ngram_count}
\title{calcal the ngrame count}
\usage{
ngram_count(word_ngram, sort = TRUE)
}
\arguments{
\item{word_ngram}{word dealed with ngrame}

\item{sort}{default true}
}
\value{
tibble
}
\description{
calcal the ngrame count
}
\examples{
ngram_count()
}
