% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/31-nse.R
\name{nse_input}
\alias{nse_input}
\title{非标准评估捕获输入参数}
\usage{
nse_input(x)
}
\arguments{
\item{x}{输入参数}
}
\value{
返回未经评估的值
}
\description{
非标准评估捕获输入参数
}
\examples{
nse_input()
nse_input(x+1)
# x+1
x <- 1
nse_input(x)
# x
}
