% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/31-nse.R
\name{nse_input_str}
\alias{nse_input_str}
\title{NSE获取输入参数的文件形式}
\usage{
nse_input_str(x)
}
\arguments{
\item{x}{参数}
}
\value{
返回值
}
\description{
NSE获取输入参数的文件形式
}
\examples{
nse_input_str()
nse_input_str(ggplot2)
get the ggplot2
nse_input_str(x+y^2)
}
