% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/31-nse.R
\name{nse_subScramble}
\alias{nse_subScramble}
\title{测试取子集后随机排序}
\usage{
nse_subScramble(x, condition)
}
\arguments{
\item{x}{数据集}

\item{condition}{条件}
}
\value{
返回值
}
\description{
测试取子集后随机排序
}
\examples{
nse_subScramble()
作为顶层函数，定义
本层用于交互，
同时为下一层提供subtitute
也就是说可以顶层才可以使用substitute
其他层全部通过表达式来传递才是函数安全的
}
