% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/31-nse.R
\name{nse_subset}
\alias{nse_subset}
\title{NSE封装过滤相关功能}
\usage{
nse_subset(x, condition)
}
\arguments{
\item{x}{数据}

\item{condition}{条件}
}
\value{
返回值
}
\description{
NSE封装过滤相关功能
}
\examples{
nse_subset()
nse_subset(iris,Species =='virginica')
总结：直接调用的话，可以不带引号
然后直接使用substitute
因为本身就是用户交互
}
