% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/31-nse.R
\name{nse_subset2}
\alias{nse_subset2}
\title{NSE封装过滤相关功能}
\usage{
nse_subset2(x, condition)
}
\arguments{
\item{x}{数据}

\item{condition}{条件}
}
\value{
返回值
}
\description{
NSE封装过滤相关功能
}
\examples{
nse_subset()
nse_subset2(iris,Species =='virginica')
虽然可以使用，但是把相关的变量从函数带到相应的相应的环境中，有点不友好
}
