% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/31-nse.R
\name{nse_subset_q}
\alias{nse_subset_q}
\title{处理子集用于函数引用}
\usage{
nse_subset_q(x, condition_q)
}
\arguments{
\item{x}{数据集}

\item{condition_q}{表达式，自带引号}
}
\value{
返回值
}
\description{
处理子集用于函数引用
}
\examples{
nse_subset_q()
如果是函数调用的话，必须形成表达式
然后才可以被eval
substitute在上一层进行定义
}
