% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/31-nse.R
\name{nse_substitute}
\alias{nse_substitute}
\title{提供一致的数据处理}
\usage{
nse_substitute(x, env = parent.frame())
}
\arguments{
\item{x}{数据}

\item{env}{环境}
}
\value{
返回值
}
\description{
提供一致的数据处理
}
\examples{
nse_subs()
a <-1
b <-2
nse_subs(a+b+2)
x<-quote(a+b)
nse_substitute(x,list(a=1,b=3))
}
