% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/31-nse.R
\name{nse_xyplot}
\alias{nse_xyplot}
\title{针对图形进行标准化处理}
\usage{
nse_xyplot(x, y, data, ...)
}
\arguments{
\item{x}{第一个参数}

\item{y}{第二个参数}

\item{data}{数据庥}

\item{...}{变量}
}
\value{
返回值
}
\description{
针对图形进行标准化处理
}
\examples{
nse_xyplot(mpg,cyl,mtcars)
nse_xyplot(mpg,hp,mtcars)
nse_xyplot(mpg,disp,mtcars,col="red",aspect="xy")
将原来写入的函数变成比较通用的方式，这是不错的处理方式
}
