% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-vec_numeric.R
\name{num_to_string_AcctFormat}
\alias{num_to_string_AcctFormat}
\title{将数字按会计格式进行处理}
\usage{
num_to_string_AcctFormat(
  x,
  divided_value = 1,
  round_digit = 2,
  seperator = ",",
  seperator_digit = 3,
  currency_symbol = ""
)
}
\arguments{
\item{x}{数据}

\item{divided_value}{如果是万元版则是10000}

\item{round_digit}{保留小数位数一般为2位小数}

\item{seperator}{千分位或万分位}

\item{seperator_digit}{千分位或万分位}

\item{currency_symbol}{币货符号}
}
\value{
返回值
}
\description{
将数字按会计格式进行处理
}
\examples{
num_to_string_AcctFormat()
num_to_string_AcctFormat(1234)
num_to_string_AcctFormat(1234.12)
num_to_string_AcctFormat(1234.123,currency_symbol = '$')
num_to_string_AcctFormat(123123123.123)
num_to_string_AcctFormat(123123123.123,divided_value = 10000)
num_to_string_AcctFormat(c(123123123.123,123321321.13),divided_value = 10000)
num_to_string_AcctFormat(c(123123123.123,123321321.13),divided_value = 10000,currency_symbol = '$')
}
