% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-vec_logical.R
\name{or_multiple}
\alias{or_multiple}
\title{扩展基本的or，用于多个变量的连接}
\usage{
or_multiple(res)
}
\arguments{
\item{res}{多个逻辑变量采用列表形式}
}
\value{
返回值
}
\description{
扩展基本的or，用于多个变量的连接
}
\examples{
or_multiple();
}
