% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-vec_character.R
\name{paging_setting}
\alias{paging_setting}
\title{设置分页处理界面}
\usage{
paging_setting(volume = 20000L, each_page = 7000L)
}
\arguments{
\item{volume}{总页数}

\item{each_page}{每页数量}
}
\value{
返回值
}
\description{
设置分页处理界面
}
\examples{
paging_setting();
}
