% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-vec_numeric.R
\name{replaceValue_lessThan_target}
\alias{replaceValue_lessThan_target}
\title{如果数值小于等于某个目标值进行替代为新值}
\usage{
replaceValue_lessThan_target(x, targetValue = 15, replaceValue = 0)
}
\arguments{
\item{x}{原始数值向量}

\item{targetValue}{目标值}

\item{replaceValue}{替代值默认为0}
}
\value{
返回值
}
\description{
如果数值小于等于某个目标值进行替代为新值
}
\examples{
replaceValue_lessThan_targetv();
}
