% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGB2CMYK.R
\name{RGB2CMYK}
\alias{RGB2CMYK}
\title{将RBG模式的颜色转移CMYK模式}
\usage{
RGB2CMYK(R, G, B)
}
\arguments{
\item{R}{红色}

\item{G}{绿色}

\item{B}{蓝字}
}
\value{
返回CMYK列表
}
\description{
将RBG模式的颜色转移CMYK模式
}
\examples{
RGB2CMYK(255,255,0);
}
