% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-color.R
\name{RGB2Hex}
\alias{RGB2Hex}
\title{用于将RGB转化为十六制定的表达方式}
\usage{
RGB2Hex(R, G, B, format = "upper", prefix = "#")
}
\arguments{
\item{R}{输入R值}

\item{G}{输入G值}

\item{B}{输入B值}

\item{format}{提供upper或lower2个选择}

\item{prefix}{使用前缀，一般为#}
}
\value{
返回值为一个数据
}
\description{
用于将RGB转化为十六制定的表达方式
}
\examples{
RGB2Hex(1:100,3:202,5:204,format = 'lower');
}
