% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-df_rowOpers.R
\name{row_del_byKeys_is}
\alias{row_del_byKeys_is}
\title{按行删除}
\usage{
row_del_byKeys_is(data, id_var, keys)
}
\arguments{
\item{data}{数据}

\item{id_var}{id}

\item{keys}{删除关键键}
}
\value{
返回值
}
\description{
按行删除
}
\examples{
row_del_byKeys_is()
}
