% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-df_rowOpers.R
\name{row_del_byKeys_like}
\alias{row_del_byKeys_like}
\title{根据行进行删除}
\usage{
row_del_byKeys_like(data, id_var, keys)
}
\arguments{
\item{data}{数据框}

\item{id_var}{id}

\item{keys}{删除关键词}
}
\value{
返回值
}
\description{
根据行进行删除
}
\examples{
row_del_byKeys_like();
}
