% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-df_rowOpers.R
\name{row_del_byNum}
\alias{row_del_byNum}
\title{按行删除按数量}
\usage{
row_del_byNum(data, id_var, num_digits)
}
\arguments{
\item{data}{数据}

\item{id_var}{ID}

\item{num_digits}{位数}
}
\value{
返回值
}
\description{
按行删除按数量
}
\examples{
row_del_byNum();
}
