% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-df_rowOpers.R
\name{row_del_byNumChars}
\alias{row_del_byNumChars}
\title{删除字母对}
\usage{
row_del_byNumChars(data, id_var, num_digits)
}
\arguments{
\item{data}{数据}

\item{id_var}{内码}

\item{num_digits}{数据位数}
}
\value{
返回值
}
\description{
删除字母对
}
\examples{
row_del_byNumChars()
}
