% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05b_time.R
\name{seconds_as_Timetext}
\alias{seconds_as_Timetext}
\title{格式化秒数}
\usage{
seconds_as_Timetext(seconds)
}
\arguments{
\item{seconds}{整数型的秒数}
}
\value{
返回值
}
\description{
格式化秒数
}
\examples{
seconds_as——TimeText()
}
