% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-vec_character.R
\name{str_contain}
\alias{str_contain}
\title{文件中是否包含字段检查}
\usage{
str_contain(x, pattern)
}
\arguments{
\item{x}{字符向量}

\item{pattern}{检查模式}
}
\value{
返回值
}
\description{
文件中是否包含字段检查
}
\examples{
str_contain(letters,'a');
}
