% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-vec_character.R
\name{str_contain_emotion}
\alias{str_contain_emotion}
\title{针对表情字段进行判断}
\usage{
str_contain_emotion(x)
}
\arguments{
\item{x}{变量}
}
\value{

}
\description{
针对表情字段进行判断
}
\examples{
str_contain_emotion()
}
