% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-vec_character.R
\name{str_contain_num}
\alias{str_contain_num}
\title{字符串包含内容}
\usage{
str_contain_num(x, digit = 11)
}
\arguments{
\item{x}{字段}

\item{digit}{数字位数}
}
\value{
返回值
}
\description{
字符串包含内容
}
\examples{
str_contain_num() 
}
