% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-vec_character.R
\name{str_contain_phone}
\alias{str_contain_phone}
\title{判断文件中是否包含电话号码}
\usage{
str_contain_phone(x)
}
\arguments{
\item{x}{向量}
}
\value{
反回值
}
\description{
判断文件中是否包含电话号码
}
\examples{
str_contain_phone()
}
