% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-vec_character.R
\name{str_contains}
\alias{str_contains}
\title{判断字符串中是否包含特定的多个字符}
\usage{
str_contains(x, keys)
}
\arguments{
\item{x}{字符串向量}

\item{keys}{多个关键词}
}
\value{
返回值
}
\description{
判断字符串中是否包含特定的多个字符
}
\examples{
str_contains(); 
}
