% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-vec_character.R
\name{str_equals}
\alias{str_equals}
\title{将字符串用于多个关键词的对比}
\usage{
str_equals(x, keys)
}
\arguments{
\item{x}{字符向量}

\item{keys}{对比关键词}
}
\value{
返回值
}
\description{
将字符串用于多个关键词的对比
}
\examples{
str_equals();
}
