% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-vec_character.R
\name{str_extractDateTime}
\alias{str_extractDateTime}
\title{获取字段日期时间部分}
\usage{
str_extractDateTime(x)
}
\arguments{
\item{x}{向量}
}
\value{
返回值
}
\description{
获取字段日期时间部分
}
\examples{
str_extractDatetime()
}
