% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-vec_character.R
\name{str_HasDateTime}
\alias{str_HasDateTime}
\title{检验字段是否包含完整的日期时间}
\usage{
str_HasDateTime(x)
}
\arguments{
\item{x}{字段}
}
\value{
返回值
}
\description{
检验字段是否包含完整的日期时间
}
\examples{
str_HasDateTime()
}
