% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-vec_character.R
\name{str_splitByLeftBrace}
\alias{str_splitByLeftBrace}
\title{将字符串分解为多列，使用英文状态下的左括号，允许前后添加内容；}
\usage{
str_splitByLeftBrace(x, prefix = "", suffix = "")
}
\arguments{
\item{x}{字符向量}

\item{prefix}{左括号前缀}

\item{suffix}{左括号后缀}
}
\value{
返回结果列表
}
\description{
将字符串分解为多列，使用英文状态下的左括号，允许前后添加内容；
}
\examples{
str_splitByLeftBrace(); 
}
