% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-vec_character.R
\name{str_splitByRighttBrace}
\alias{str_splitByRighttBrace}
\title{将字符串进行拆分为多列，使用右括号}
\usage{
str_splitByRighttBrace(x, prefix = "", suffix = "")
}
\arguments{
\item{x}{字符向量}

\item{prefix}{右括号前缀}

\item{suffix}{右括号后缀}
}
\value{
返回值
}
\description{
将字符串进行拆分为多列，使用右括号
}
\examples{
str_splitByRighttBrace();
}
