% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-vec_character.R
\name{str_splitCols}
\alias{str_splitCols}
\title{将一列数据分解为多列数据,返回列表}
\usage{
str_splitCols(x, sep = ",")
}
\arguments{
\item{x}{将要处理的处理}

\item{sep}{分隔符}
}
\value{
返回值
}
\description{
将一列数据分解为多列数据,返回列表
}
\examples{
str_splitCols();
}
