% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-vec_character.R
\name{str_suffix_complete}
\alias{str_suffix_complete}
\title{自动补全功能}
\usage{
str_suffix_complete(x, suffix = "/")
}
\arguments{
\item{x}{字符串}

\item{suffix}{后缀}
}
\value{
返回值
}
\description{
自动补全功能
}
\examples{
str_suffix_complete()
}
