% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-vec_numeric.R
\name{string_acctFormat_to_num}
\alias{string_acctFormat_to_num}
\title{将会计格式金额格式化为数值}
\usage{
string_acctFormat_to_num(x, seperator = ",", currency_symbol = "")
}
\arguments{
\item{x}{会计格式的数值}

\item{seperator}{千分位}

\item{currency_symbol}{货币符号}
}
\value{
返回数值
}
\description{
将会计格式金额格式化为数值
}
\examples{
string_acctFormat_to_num()
string_acctFormat_to_num('$123,123,123.12',currency_symbol = '$')
string_acctFormat_to_num('￥123,123,123.12',currency_symbol = '￥')

}
