% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-vec_character.R
\name{text_dcast}
\alias{text_dcast}
\title{将数据框按行按行进行扩展}
\usage{
text_dcast(data, var_rows = c("FNumber", "FName"), var_cols = "FType",
  var_values = c("FSPID", "FSPName"))
}
\arguments{
\item{data}{数据}

\item{var_rows}{行上数据}

\item{var_cols}{列上数据}

\item{var_values}{显示数据}
}
\value{
返回值
}
\description{
将数据框按行按行进行扩展
}
\examples{
text_dcast();
}
