% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-vec_character.R
\name{text_prod}
\alias{text_prod}
\title{处理两个字符向量的笛卡尔积展开}
\usage{
text_prod(x, y, sep = "#")
}
\arguments{
\item{x}{第一个向量}

\item{y}{第二个向量}

\item{sep}{分隔符}
}
\value{
返回向量
}
\description{
处理两个字符向量的笛卡尔积展开
}
\examples{
text_prod(letters,LETTERS);
}
