% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05b_time.R
\name{time_formatter}
\alias{time_formatter}
\title{格式化时间部分}
\usage{
time_formatter(value = "0", unit = "年")
}
\arguments{
\item{value}{值}

\item{unit}{单位}
}
\value{
返回值
}
\description{
格式化时间部分
}
\examples{
time_formatter()
}
