% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-vec_character.R
\name{vect_as_dbl_equal}
\alias{vect_as_dbl_equal}
\title{处理字符串连接问题}
\usage{
vect_as_dbl_equal(x)
}
\arguments{
\item{x}{字符赂量}
}
\value{
返回值
}
\description{
处理字符串连接问题
}
\examples{
vect_as_dbl_equal();
}
