% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-vec_character.R
\name{vect_as_list}
\alias{vect_as_list}
\title{处理将向量的每个元素装入一个列表}
\usage{
vect_as_list(x)
}
\arguments{
\item{x}{原始向量}
}
\value{
返回值
}
\description{
处理将向量的每个元素装入一个列表
}
\examples{
vect_as_list(letters);
}
