% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-vec_character.R
\name{vect_as_long_string}
\alias{vect_as_long_string}
\title{将向量转化为长字符串}
\usage{
vect_as_long_string(x)
}
\arguments{
\item{x}{向量}
}
\value{
返回值
}
\description{
将向量转化为长字符串
}
\examples{
vect_as_long_string();
}
