% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_word.R
\name{vector_col_del}
\alias{vector_col_del}
\title{处理替代数据}
\usage{
vector_col_del(x, keyWord)
}
\arguments{
\item{x}{原始数据}

\item{keyWord}{关键词}
}
\value{
返回值
}
\description{
处理替代数据
}
\examples{
vector_col_del()
}
