% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_word.R
\name{vector_col_dels}
\alias{vector_col_dels}
\title{删除相应的词逻辑}
\usage{
vector_col_dels(x, keyWords)
}
\arguments{
\item{x}{原始词}

\item{keyWords}{关键词}
}
\value{
返回值
}
\description{
删除相应的词逻辑
}
\examples{
vector_col_dels();
vector_col_dels(letters,letters[1:10]);
}
