% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_word.R
\name{vector_del_rightBrace}
\alias{vector_del_rightBrace}
\title{删除右边空格}
\usage{
vector_del_rightBrace(x)
}
\arguments{
\item{x}{向量}
}
\value{
返回值
}
\description{
删除右边空格
}
\examples{
vector_del_rightBrace();
}
