% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-vec_character.R
\name{vector_not_in}
\alias{vector_not_in}
\title{处理向量不包含的问题}
\usage{
vector_not_in(all, part)
}
\arguments{
\item{all}{全部}

\item{part}{部分}
}
\value{
返回值
}
\description{
处理向量不包含的问题
}
\examples{
vector_not_in();
}
