% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_word.R
\name{word_col_del}
\alias{word_col_del}
\title{针对数据处理删词处理}
\usage{
word_col_del(data, value_var, keyWords)
}
\arguments{
\item{data}{数据框}

\item{value_var}{指定变量}

\item{keyWords}{关键词}
}
\value{
返回值
}
\description{
针对数据处理删词处理
}
\examples{
word_col_del();
library(tsda);
mydata <- nsim_read('item_question');
View(mydata);
mydata2 <- word_col_del(mydata,'FQuestion',c('[图片]','[卡片]'));
View(mydata2);
}
