% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_textmining.R
\name{word_count_plot}
\alias{word_count_plot}
\title{plot the word count}
\usage{
word_count_plot(word_count_tidy, miss_count = 100, xlab = NULL)
}
\arguments{
\item{word_count_tidy}{word tidyed}

\item{miss_count}{defalut 100}
}
\value{
return valued with tibble
}
\description{
plot the word count
}
\examples{
word_count_plot
}
