% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_textmining.R
\name{word_count_plot_cloud}
\alias{word_count_plot_cloud}
\title{plot the wordcount with wordcloud plot}
\usage{
word_count_plot_cloud(word_count_tidy, max_words = 100)
}
\arguments{
\item{word_count_tidy}{wordcount}

\item{max_words}{show the maximun words}
}
\value{
plot
}
\description{
plot the wordcount with wordcloud plot
}
\examples{
word_count_plot_cloud()
}
