% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_textmining.R
\name{word_count_tidy}
\alias{word_count_tidy}
\title{count the word with order}
\usage{
word_count_tidy(tidy_books, sort = TRUE)
}
\arguments{
\item{tidy_books}{books tidied}

\item{sort}{defalut TRUE}
}
\value{
tibble
}
\description{
count the word with order
}
\examples{
word_count_dity()
}
