% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_textmining.R
\name{word_ngram}
\alias{word_ngram}
\title{split the word into ngram}
\usage{
word_ngram(books, ncount = 3)
}
\arguments{
\item{books}{books contain text col}

\item{ncount}{count of the word}
}
\value{
tribble
}
\description{
split the word into ngram
}
\examples{
word_ngram()
}
